Fix showUploadSpecimens link. Move specimen events action.
