+++ study/src/org/labkey/study/view/specimen/specimen.jsp
@@ -17,7 +17,8 @@
 %>
 <%@ page import="org.labkey.api.specimen.SpecimenManager"%>
 <%@ page import="org.labkey.api.specimen.Vial"%>
-<%@ page import="org.labkey.api.specimen.location.LocationManager"%>
+<%@ page import="org.labkey.api.specimen.actions.SpecimenEventBean"%>
+<%@ page import="org.labkey.api.specimen.location.LocationManager" %>
 <%@ page import="org.labkey.api.specimen.model.SpecimenComment" %>
 <%@ page import="org.labkey.api.specimen.security.permissions.SetSpecimenCommentsPermission" %>
 <%@ page import="org.labkey.api.study.Location" %>
@@ -25,7 +26,6 @@
 <%@ page import="org.labkey.api.view.ActionURL" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
-<%@ page import="org.labkey.study.controllers.specimen.SpecimenController.SpecimenEventBean" %>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.UpdateCommentsAction" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
